+++ src/gui.py
@@ -23,10 +23,12 @@ def disc():
     msg = StringVar()  # pour le message qui sera envoyé
     label = Label(wd, text=tr("status.connected"), font=("courrier", 22), bg="#56646A", fg="white")
     label.pack(side=TOP)
-    msg_list = Text(wd, bg="#545454", height=30, width=100, state=DISABLED)
-    msg_list.pack()
+    msg_list = Text(wd, bg="#545454", fg="white", state=DISABLED)
+    msg_list.pack(expand=True, fill='both', padx=100)
     chp = Entry(wd, width=70, font=(22), bg="#56646A", fg="white", bd=2, relief=SUNKEN, textvariable=msg)
-    chp.pack(side=BOTTOM, pady=10)
+    chp.insert(0, tr("gui.message.placeholder"))
+    chp.pack(side=BOTTOM, pady=10, padx=100, fill="x")
+    chp.bind("<FocusIn>", lambda args: chp.delete('0', 'end'))
 
     def send(e):
         Thread(target=interface.parseCommand, args=[msg.get()]).start()