+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Placeholder in GUI
+
+### Changed
+- GUI messages now in white
+
 ## [2.3.0] - 2020-05-29
 ### Added
 - File uploading and saving
@@ -264,6 +271,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic server and client
 - Use regex to parse commands
 
+[Unreleased]: https://github.com/RomainL972/isn/compare/v2.3.0...HEAD
 [2.3.0]: https://github.com/RomainL972/isn/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/RomainL972/isn/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/RomainL972/isn/compare/v2.0.0...v2.1.0